requirements.txt
boto3
json-schema-matcher
psycopg2-binary